+++ frontend/src/App.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import SpaceBackground from "./components/SpaceBackground/SpaceBackground";
 import Header from "./components/Header/Header";
 import IndexPage from "./pages/IndexPage";
-import LoginPage from "./pages/LoginPage";
+import LoginPage from "./pages/LoginPage/LoginPage";
 import RegisterPage from "./pages/RegisterPage/RegisterPage";
 
 import axios from "./api";
