+++ frontend/src/components/UI/Button/Button.module.css
@@ -49,9 +49,8 @@
   font-size: 2.5rem;
 }
 
-.submitBtn {
+.btnWithLoader {
   min-height: 3.5rem;
-  margin-top: 1.8rem;
 
   display: flex;
   justify-content: center;
