+++ frontend/src/pages/LoginPage/LoginPage.module.css
@@ -1,0 +1,3 @@
+.submitBtn {
+  margin-top: 1.2rem;
+}
