+++ frontend/src/pages/RegisterPage/RegisterPage.module.css
@@ -1,0 +1,3 @@
+.submitBtn {
+  margin-top: 1.6rem;
+}
